Move renderer to asciimatics
